fix: don't attempt to parse message timestamps
On the day DST "falls back", OCS timestamps during the 1am-2am hour are
ambiguous, as they don't specify _which_ of the two occurrences of this
hour they are in. We could make an educated guess, but currently the
only consumer of these events (RTR) ignores all fields except `raw`
anyway, so the simplest solution is to not try to populate `time` from
the messages at all.

Since this would mean `time` and `data.received_time` are always
identical, we also remove the latter field. Normally we'd consider this
a backwards-incompatible change and introduce a new value of `type`,
but since currently the only consumer of this data ignores the field,
and we're not storing the events anywhere, there are no compatibility
concerns.