refactor: only run a single HealthChecker
It's not a good idea to run unsupervised processes: the Supervisor
started by the Proxy was one of these cases. It also resulted in
duplication of the `:ranch.info/1` logs, as they would be logged for
each open connection.

This changes the HealthChecker to be a single process, which uses
`:ranch.procs/2` to get the list of Proxy processes running and logs
their mailbox size that way.

We can also use the built-in `:erlang.process_info/2` function to avoid
the dependency on `recon`.